Minor cleanup to DRY up duplicate code
