+++ src/payment/mod.rs
@@ -14,7 +14,6 @@ mod spontaneous;
 pub(crate) mod store;
 mod unified_qr;
 
-pub use bolt11::Bolt11InvoiceDescription;
 pub use bolt11::Bolt11Payment;
 pub use bolt12::Bolt12Payment;
 pub use onchain::OnchainPayment;
